addr: Fixed address denormalization not taking segment gaps into account
The old code assumed that image contents were loaded contiguously. This
may be false in practice, when for instance the executable segment is
loaded at a different base than the data segment, leaving a random gap
in between (courtesy of kernel ASLR).

We fix this behaviour by taking the segment index into account when
normalizing and denormalizing addresses.

Furthermore, this commit now makes the ELF code check for the image
boundaries. Before that, due to a limitation of the link_map interface,
it only compared the image base with the address and supposed that the
closest lower base was the image the address belonged to.

Fixes Snaipe/Criterion#178.

Signed-off-by: Franklin "Snaipe" Mathieu <snaipe@diacritic.io>