arena: fixed the page mapping check
The old code was based on msync(addr, page_size, 0), which tested
whether a page was mapped at addr or not. This behaviour however was
undocumented and fragile; and eventually broke on BashForWindows.

The new method now uses mincore(2) to achieve a similar functionality,
without resorting to passing MS_(A)SYNC to msync and potentially having
side effects.