asm: make stack non-executable
This patch makes boxfort stop asking for an executable stack, which is
the default when compiling assembly bits. This was not too much of a
problem previously, but some platforms, like WSL, does not support
executable stacks yet, which forced users to clear the executable stack
bit in the generated ELFs.