exe,elf: fix main symbol resolution
The dynamic symbol table search for main yielded the address of the PLT
stub on clang, which in turn is not used by the entry point. This was
causing Criterion to fork-bomb the user when running test workers, as
they would spawn a new runner every time.

Users are advised to use a check mechanism, like setting an environment
variable, and abort immediately when seeing this environment variable
in the runner's main function.