arena: fix out-of-bounds access from the arena chunk resize.
This is an improvement over 831428f. Instead of initializing a sentinel
at the end of the arena, we simply check that there is enough space for
the next chunk to exist. If the arena is resized we perform a new search
for the best free chunk. It's not ideal but it's simple.