meson: force static lib build
boxfort can't really function well as a shared lib, and meson
currently just doesn't know how to handle a shared boxfort as a
subproject. In particular, boxfort **needs** b_lundef=false set,
but when used as a subproject it inherits the parent's b_lundef,
which is half the time completely wrong.